refactor: resolve remaining directory-architecture findings
- functions/ を削除。Workers 構成（main = src/worker.ts）では読まれず、
  CI・wrangler・scripts・vite のいずれからも参照されていなかった。
  稼働している contact API は src/routes/api/contact.ts。
- テスト配置を統一。モジュール単位のテストは対象の隣の __tests__/ に置き、
  src/tests/ は wrangler/CI/vite を横断するリポジトリ全体の検証のみとする。
- createServerFn モジュール（highlight/markdown）を src/lib/server/ に隔離し、
  クライアント用ユーティリティとのサーバー境界を明示。
  src/lib/microcms.ts は後方互換ファサードとしてパス自体が契約のため据え置く。
- 刷新で陳腐化していた src/server/markdown/home.ts を現在のサイト内容に更新
  （Nanatau / WebGL / Next.js の記述を削除）。
- CLAUDE.md にディレクトリツリーとテスト配置規約を反映。
