+++ src/routes/__tests__/privacy-policy-layout.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from 'bun:test'
 import { readFileSync } from 'node:fs'
 import { resolve } from 'node:path'
 
-const privacyPolicyRoutePath = resolve(import.meta.dir, '../routes/privacy-policy.tsx')
+const privacyPolicyRoutePath = resolve(import.meta.dir, '../privacy-policy.tsx')
 const privacyPolicySource = readFileSync(privacyPolicyRoutePath, 'utf8')
 
 describe('privacy-policy route layout', () => {
