+++ src/routes/article/$slug.tsx
@@ -4,10 +4,10 @@ import CodeHighlight from '@/components/article/CodeHighlight'
 import Breadcrumb from '@/components/layout/Breadcrumb'
 import SitePage from '@/components/layout/SitePage'
 import JsonLd, { createArticleSchema, createBreadcrumbSchema } from '@/components/seo/JsonLd'
-import { highlightContent } from '@/lib/highlight'
-import { parseContentMarkdown } from '@/lib/markdown'
 import { getBlogDetail } from '@/lib/microcms'
 import { getSeoDescription, getSeoMetadata, getSeoTitle } from '@/lib/seoMetadata'
+import { highlightContent } from '@/lib/server/highlight'
+import { parseContentMarkdown } from '@/lib/server/markdown'
 import { createStandardHead, DEFAULT_OG_IMAGE_URL, toCanonicalUrl } from '@/lib/siteMetadata'
 import { formatDate } from '@/lib/utils'
 import {
